Worklet: Implement exception report path for ThreadedWorklet
This CL implements exception report path for ThreadedWorklet. Specifically, this
exposes MessageHandlerInWorker() that is a file-scoped static function to
V8Initializer, and makes V8ScriptRunner::ReportException() call it when it's on
a worker thread.

For confirming this fix, this CL adds WPT tests for a syntax error and an
invalid module specifier on worklets, and fixes error handling in
WorkletModuleTreeClient::NotifyModuleTreeLoadFinished().

Bug: 781300
Change-Id: Id9be208ecbc594b4e33cb07279f5b987107289e5